fix some known errors

+++ dbms/src/Storages/DeltaMerge/DMVersionFilterBlockInputStream.h
@@ -62,11 +62,11 @@ public:
         LOG_FMT_DEBUG(log,
                       "Total rows: {}, pass: {:.2f}%, complete pass: {:.2f}%, complete not pass: {:.2f}%, not clean: {:.2f}%, effective: {:.2f}%, read tso: {}, query id: {}",
                       total_rows,
-                      (Float64)passed_rows * 100 / total_rows,
-                      (Float64)complete_passed * 100 / total_blocks,
-                      (Float64)complete_not_passed * 100 / total_blocks,
-                      (Float64)not_clean_rows * 100 / passed_rows,
-                      (Float64)effective_num_rows * 100 / passed_rows,
+                      passed_rows * 100.0 / total_rows,
+                      complete_passed * 100.0 / total_blocks,
+                      complete_not_passed * 100.0 / total_blocks,
+                      not_clean_rows * 100.0 / passed_rows,
+                      effective_num_rows * 100.0 / passed_rows,
                       version_limit,
                       (query_id.empty() ? "<non-query>" : query_id));
     }
